Reorganized code
- Code from beckett_tweaks.py is now in api.py and resources_v2.py
- Removed beckett_tweaks.py
